requirements.txt
alembic==1.4.3
arabic-reshaper==2.1.3
asgiref==3.5.2
Babel==2.8.0
backports.zoneinfo==0.2.1;python_version<"3.9"
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
certifi==2021.10.8
cffi==1.14.3
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-static==0.0.6
Django==3.2
django-ckeditor==6.0.0
django-filter==2.4.0
django-import-export==2.3.0
django-js-asset==1.2.2
dnspython==2.0.0
email-validator==1.1.1
et-xmlfile==1.0.1
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
ldap3==2.8.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==1.1.1
MouseInfo==0.1.3
odfpy==1.4.1
openpyxl==3.0.5
paramiko==2.7.2
passlib==1.7.2
pgadmin4==4.26
Pillow==8.0.1
psutil==5.7.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.4.0
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-bidi==0.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytweening==1.0.4
pytz==2018.9
pywhatkit==5.3
PyYAML==5.3.1
reportlab==3.6.1
requests==2.27.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.3.20
sqlparse==0.2.4
sshtunnel==0.1.5
static3==0.7.0
style==1.1.0
tablib==2.0.0
tzdata==2022.2
update==0.0.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
